+++ app/src/main/java/unibo/cineradar/model/review/FilmReview.java
@@ -22,7 +22,7 @@ public class FilmReview extends Review {
                       final String username,
                       final String title,
                       final String description,
-                      final int overallRating) {
+                      final double overallRating) {
         super(username, title, description, overallRating);
         this.idFilm = idFilm;
         this.filmTitle = filmTitle;
